src/main/java/org/sonar/plugins/scmactivity/sha1/Sha1Generator.java
/*
 * Sonar SCM Activity Plugin
 * Copyright (C) 2010 SonarSource
 * dev@sonar.codehaus.org
 *
 * This program is free software; you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public
 * License as published by the Free Software Foundation; either
 * version 3 of the License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public
 * License along with this program; if not, write to the Free Software
 * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02
 */

package org.sonar.plugins.scmactivity.sha1;

import com.google.common.io.Closeables;

import org.apache.commons.codec.digest.DigestUtils;
import org.sonar.api.BatchExtension;

import java.io.BufferedInputStream;
import java.io.File;
import java.io.FileInputStream;
import java.io.IOException;
import java.io.InputStream;

public class Sha1Generator implements BatchExtension {
  public String sha1(File file) throws IOException {
    InputStream input = null;
    try {
      input = new BufferedInputStream(new FileInputStream(file));
      return DigestUtils.shaHex(input);
    } finally {
      Closeables.closeQuietly(input);
    }
  }
}
